refactor: split editor reminder pill display into modular CM6 components and remove legacy file
- Delete monolithic src/plugin/ui/editor-reminder-display.ts
- Add modular CM6 components under src/plugin/ui/editor-reminder-display/:
  - index.ts for exports
  - pill-widget.ts for widget rendering and activation
  - spans.ts for span derivation helpers
  - styles.ts for idempotent CSS injection
  - state-effects.ts for recompute effect
  - types.ts for shared types
- Improves maintainability and separation of concerns without changing behavior
