+++ src/plugin/ui/editor-reminder-display/index.ts
@@ -1,0 +1,4 @@
+export { createReminderPillExtension } from "./extension";
+export { ensureReminderPillStylesInjected } from "./styles";
+export { forceReminderPillRecompute } from "./state-effects";
+export type { TokenSpan, PillSpec, PillContext } from "./types";
